requirements.txt
Flask==2.2.2
urllib3==1.26.12
Pillow==9.2.0
timm==0.6.7
gunicorn
--extra-index-url https://download.pytorch.org/whl/cpu
torch